[Chore] Restructure folder setup
